Stopwatch Web Application
Stopwatch Web Application
This is a simple web application that allows you to track time using a stopwatch. It can be used to time events or activities, and it's perfect for use in sports, cooking, or any other activity where you need to measure elapsed time.

Screenshot of Stopwatch Web Application

Features
Start, stop, and reset the stopwatch
Track elapsed time with precision up to 1/1000th of a second
Lap functionality to record split times
Technologies Used
This web application was built using the following technologies:

HTML
CSS
JavaScript
How to Use
To use the stopwatch web application, simply follow these steps:

Clone this repository to your local machine.
Open index.html in your web browser.
Click the "Start" button to begin timing.
Click the "Stop" button to stop timing.
Click the "Reset" button to reset the stopwatch to 0.
Click the "Lap" button to record a split time.
Contributing
Contributions to this project are welcome. To contribute, simply fork this repository, make your changes, and submit a pull request.

License
This project is licensed under the MIT License. See the LICENSE file for more information.